+++ run.sh
@@ -196,6 +196,8 @@ run_sandbox() {
   # protocol_runner needs 'libtezos.so' to run
   export LD_LIBRARY_PATH="${BASH_SOURCE%/*}/tezos/interop/lib_tezos/artifacts:${BASH_SOURCE%/*}/target/$PROFILE"
 
+  export TEZOS_CLIENT_UNSAFE_DISABLE_DISCLAIMER="Y"
+
   rm -rf ./light_node/etc/tezedge_sandbox/tezos-client/*
 
   cargo run $CARGO_PROFILE_ARG --bin sandbox -- \