+++ CHANGELOG.md
@@ -31,6 +31,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Nothing.
 
 
+## [0.4.0] - 2020-09-16
+
+### Added
+
+- More verbose error handling in the sandbox launcher.
+- New rpc `forge/operations`.
+- New docker-compose file to start a setup with the sandbox launcher, tezedge-explorer front-end and tezedge-debugger.
+
+### Changed
+
+- Nothing.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Various bugs in the sandbox launcher.
+
+### Security
+
+- Nothing.
+
+
 ## [0.3.0] - 2020-08-31
 
 ### Added
@@ -140,12 +169,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to connect and bootstrap data from Tezos Babylonnet.
 - Protocol FFI integration.
 
-[Unreleased]: https://github.com/simplestaking/tezedge/compare/v0.0.3...HEAD
+[Unreleased]: https://github.com/simplestaking/tezedge/compare/v0.0.4...HEAD
 [0.0.1]: https://github.com/simplestaking/tezedge/releases/v0.0.1
 [0.0.2]: https://github.com/simplestaking/tezedge/releases/v0.0.2
 [0.1.0]: https://github.com/simplestaking/tezedge/releases/v0.1.0
 [0.2.0]: https://github.com/simplestaking/tezedge/releases/v0.2.0
 [0.3.0]: https://github.com/simplestaking/tezedge/releases/v0.3.0
+[0.4.0]: https://github.com/simplestaking/tezedge/releases/v0.4.0
 ___
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).